Ajustes en el api: tipos de datos y configuraciones

+++ Services/TeamTasksApi/Domain/Views/DeveloperLoadSummaryView.cs
@@ -4,6 +4,6 @@
     {
         public string DeveloperName { get; set; }
         public int OpenTasksCount { get; set; }
-        public double AverageEstimatedComplexity { get; set; }
+        public int AverageEstimatedComplexity { get; set; }
     }
 }
